Fix shutdown of server process
