+++ crates/remote_server/src/remote_server.rs
@@ -1,5 +1,8 @@
 mod headless_project;
 
+#[cfg(not(windows))]
+pub mod unix;
+
 #[cfg(test)]
 mod remote_editing_tests;
